[Refactor] userhome 디렉토리 정리
