+++ src/main/java/com/hongik/devtalk/repository/userhome/SeminarInfoRepository.java
@@ -1,4 +1,4 @@
-package com.hongik.devtalk.repository;
+package com.hongik.devtalk.repository.userhome;
 
 
 import com.hongik.devtalk.domain.Seminar;
